# sira-backend-master 2/swagger-api.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/settings": {
            "post": {
                "operationId": "SettingsController_create",
                "summary": "Create a new setting",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSettingDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Settings"
                ]
            },
            "get": {
                "operationId": "SettingsController_findAll",
                "summary": "Get all settings",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Settings"
                ]
            }
        },
        "/settings/key/{key}": {
            "get": {
                "operationId": "SettingsController_findByKey",
                "summary": "Get a setting by its unique key",
                "parameters": [
                    {
                        "name": "key",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Settings"
                ]
            }
        },
        "/settings/{id}": {
            "get": {
                "operationId": "SettingsController_findOne",
                "summary": "Get a setting by ID",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Settings"
                ]
            },
            "patch": {
                "operationId": "SettingsController_update",
                "summary": "Update a setting",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSettingDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Settings"
                ]
            },
            "delete": {
                "operationId": "SettingsController_remove",
                "summary": "Delete a setting",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Settings"
                ]
            }
        },
...
    },
    "info": {
        "title": "CRM API",
        "description": "Comprehensive CRM API for real estate management",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateSettingDto": {
                "type": "object",
                "properties": {
                    "settingKey": {
                        "type": "string"
                    },
                    "settingValue": {
                        "type": "string"
                    },
                    "dataType": {
                        "type": "string",
                        "enum": [
                            "string",
                            "integer",
                            "boolean",
                            "json"
                        ],
                        "default": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "settingKey"
                ]
            },
            "UpdateSettingDto": {
                "type": "object",
                "properties": {
                    "settingKey": {
                        "type": "string"
                    },
                    "settingValue": {
                        "type": "string"
                    },
                    "dataType": {
                        "type": "string",
                        "enum": [
                            "string",
                            "integer",
                            "boolean",
                            "json"
                        ],
                        "default": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}